% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleSnippets-package.r
\docType{package}
\name{NimbleSnippets}
\alias{NimbleSnippets}
\alias{NimbleSnippets-package}
\title{Nimble Snippets}
\description{
A collection of nimble functions.
}
\section{Adaptive parallel tempering (APT)}{
 Funtions for adaptive parallel tempering are provided. These were adapted from nimble's MCMC suite to enable APT specific features (i.e. an adaptive temperature ladder) to be included.
}

\seealso{
\url{https://r-nimble.org/}
}
