% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APT_functions.R
\name{plotTempTraj}
\alias{plotTempTraj}
\title{plot.tempTraj}
\usage{
plotTempTraj(cAPT)
}
\arguments{
\item{cAPT}{- an APT object generated by buildAPT and compiled by compileNimble.}
}
\value{
A plot of the trajectories.
}
\description{
Plot the trajectories of a temperature ladder of an adaptive parallel tempering algorithm
}
\details{
Returns two plots, one with T~iterations, the other log10(T)~iterations.
}
\author{
David Pleydell
}
\seealso{
An example is provided in the documentation of buildAPT.
}

